Remove `SourceRole.Unknown` and use nullability instead. Rename `Cond(Not)NullEffect` to `Returns(Not)NullEffect`
See comments:
* https://github.com/jesyspa/kotlin/pull/159#discussion_r1376032386
* https://github.com/jesyspa/kotlin/pull/159#discussion_r1376031186
* https://github.com/jesyspa/kotlin/pull/159#discussion_r1372981031
