+++ plugins/formal-verification/formver.core/src/org/jetbrains/kotlin/formver/embeddings/expression/Comparison.kt
@@ -5,10 +5,10 @@
 
 package org.jetbrains.kotlin.formver.embeddings.expression
 
-import org.jetbrains.kotlin.formver.asInfo
 import org.jetbrains.kotlin.formver.asPosition
 import org.jetbrains.kotlin.formver.embeddings.BooleanTypeEmbedding
 import org.jetbrains.kotlin.formver.info.SourceRole
+import org.jetbrains.kotlin.formver.info.asInfo
 import org.jetbrains.kotlin.formver.linearization.LinearizationContext
 import org.jetbrains.kotlin.formver.viper.ast.Exp
 
@@ -50,7 +50,7 @@ data class GeCmp(
 data class EqCmp(
     override val left: ExpEmbedding,
     override val right: ExpEmbedding,
-    val sourceRole: SourceRole = SourceRole.Unknown,
+    val sourceRole: SourceRole? = null,
 ) : ComparisonExpression {
     override fun toViper(ctx: LinearizationContext) =
         Exp.EqCmp(left.toViper(ctx), right.toViper(ctx), ctx.source.asPosition, sourceRole.asInfo)
@@ -59,7 +59,7 @@ data class EqCmp(
 data class NeCmp(
     override val left: ExpEmbedding,
     override val right: ExpEmbedding,
-    val sourceRole: SourceRole = SourceRole.Unknown,
+    val sourceRole: SourceRole? = null,
 ) : ComparisonExpression {
     override fun toViper(ctx: LinearizationContext) =
         Exp.NeCmp(left.toViper(ctx), right.toViper(ctx), ctx.source.asPosition, sourceRole.asInfo)
