+++ plugins/formal-verification/formver.core/src/org/jetbrains/kotlin/formver/embeddings/expression/Literal.kt
@@ -6,7 +6,6 @@
 package org.jetbrains.kotlin.formver.embeddings.expression
 
 import org.jetbrains.kotlin.KtSourceElement
-import org.jetbrains.kotlin.formver.asInfo
 import org.jetbrains.kotlin.formver.asPosition
 import org.jetbrains.kotlin.formver.domains.NullableDomain
 import org.jetbrains.kotlin.formver.embeddings.BooleanTypeEmbedding
@@ -14,6 +13,7 @@ import org.jetbrains.kotlin.formver.embeddings.IntTypeEmbedding
 import org.jetbrains.kotlin.formver.embeddings.NullableTypeEmbedding
 import org.jetbrains.kotlin.formver.embeddings.TypeEmbedding
 import org.jetbrains.kotlin.formver.info.SourceRole
+import org.jetbrains.kotlin.formver.info.asInfo
 import org.jetbrains.kotlin.formver.linearization.LinearizationContext
 import org.jetbrains.kotlin.formver.viper.ast.Exp
 
@@ -28,7 +28,7 @@ data class IntLit(val value: Int) : PureExpEmbedding {
     override fun toViper(source: KtSourceElement?): Exp = Exp.IntLit(value, source.asPosition)
 }
 
-data class BooleanLit(val value: Boolean, val sourceRole: SourceRole = SourceRole.Unknown) : PureExpEmbedding {
+data class BooleanLit(val value: Boolean, val sourceRole: SourceRole? = null) : PureExpEmbedding {
     override val type = BooleanTypeEmbedding
 
     override fun toViper(source: KtSourceElement?): Exp = Exp.BoolLit(value, source.asPosition, sourceRole.asInfo)
